+++ PolyEngine/RenderingDevice/OpenGL/Src/ParticlesRenderingPass.cpp
@@ -55,7 +55,7 @@ void ParticlesRenderingPass::OnRun(World* world, const CameraComponent* camera, 
 		GetProgram().SetUniform("uSpriteSpeed", spriteSettings.Speed);
 		GetProgram().SetUniform("uSpriteSubImages", spriteSettings.SubImages.X, spriteSettings.SubImages.Y);
 
-		int partileLen = particleCmp->GetEmitter()->GetInstancesCount();
+		int partileLen = (int)(particleCmp->GetEmitter()->GetInstancesCount());
 		const TextureResource* Texture = particleCmp->GetEmitter()->GetSpritesheet();
 		const GLParticleDeviceProxy* particleProxy = static_cast<const GLParticleDeviceProxy*>(particleCmp->GetEmitter()->GetParticleProxy());
 		GLuint particleVAO = particleProxy->GetVAO();
